always materalize volitile functions AND distinguish CTEs in logical plan to be optimized in physical plan
