+++ datafusion/optimizer/src/lib.rs
@@ -59,6 +59,7 @@ pub mod eliminate_outer_join;
 pub mod extract_equijoin_predicate;
 pub mod extract_leaf_expressions;
 pub mod filter_null_join_keys;
+pub mod materialize_cte;
 pub mod optimize_projections;
 pub mod optimize_unions;
 pub mod optimizer;
